requirements.txt
cffi==2.0.0
cryptography==46.0.5
joblib==1.5.3
numpy==2.4.2
pandas==3.0.0
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
